fix loop
